Created components for login, signup and review -abstracted code to components and removed it from the rest of the code.
