+++ client/src/components/review/movieInfo.tsx
@@ -1,16 +1,7 @@
 import * as React from "react";
 import NoPoster from "./NoPoster.jpg";
 import "./style.css";
-import { ILoginState, IState, IMovieState } from "../../redux/reducers";
-import {
-  loginEmailUpdate,
-  loginPasswordUpdate,
-  signupPasswordUpdate,
-  signupEmailUpdate,
-  signupConfirmPassword,
-  signupFirstnameUpdate,
-  signupLastnameUpdate
-} from "../../redux/actions/loginsignup.actions";
+import { IState, IMovieState } from "../../redux/reducers";
 import {
   movieSearchResolved,
   userSubmitRequest,
@@ -18,9 +9,11 @@ import {
   toDashboard
 } from "../../redux/actions/movie.actions";
 import { connect } from "react-redux";
+import Signup from "../Modals/Signup";
+import Login from "../Modals/Login";
+import Review from "../Modals/Review";
 
 export interface IMovieInfoProps {
-  login: ILoginState;
   movie: IMovieState;
 
   // Action properties from movie actions
@@ -33,79 +26,9 @@ export interface IMovieInfoProps {
   userSubmitRequest: () => void;
   inputUpdate: (inputValue: string) => void;
   toDashboard: (toDashboard: boolean) => void;
-
-  // action porperties from login actions
-  loginEmailUpdate: (email: string) => void;
-  loginPasswordUpdate: (password: string) => void;
-  signupEmailUpdate: (email: string) => void;
-  signupPasswordUpdate: (password: string) => void;
-  signupConfirmPassword: (password: string) => void;
-  signupFirstnameUpdate: (fn: string) => void;
-  signupLastnameUpdate: (ln: string) => void;
 }
 
 class MovieInfo extends React.Component<IMovieInfoProps> {
-  handleLoginEmailUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.loginEmailUpdate(value);
-  }
-
-  handleLoginPasswordUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.loginPasswordUpdate(value);
-  }
-
-  handleSignupEmailUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.signupEmailUpdate(value);
-  }
-
-  handleSignupPasswordUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.signupPasswordUpdate(value);
-  }
-
-  handleSignupConfirmPassword(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.signupConfirmPassword(value);
-  }
-
-  handleSignupFirstnameUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.signupFirstnameUpdate(value);
-  }
-
-  handleSignupLastnameUpdate(e: any) {
-    console.log("input changing");
-    const value = e.target.value;
-    this.props.signupLastnameUpdate(value);
-  }
-
-  signupUser() {
-    console.log(this.props.login.firstname);
-    console.log(this.props.login.lastname);
-    console.log(this.props.login.signupEmail);
-    console.log(this.props.login.signupPassword);
-    console.log(this.props.login.confirmPassword);
-  }
-
-  /**
-   * Contains the axios call to the backend to log in a user using
-   * the email and password they type in the login modal
-   * @param em email used to login user
-   * @param pw password used to login a user
-   */
-  loginUser() {
-    console.log(this.props.login);
-    console.log(this.props.login.loginEmail);
-    console.log(this.props.login.loginPassword);
-  }
 
   getPoster(): string {
     if (this.props.movie.posterUrl === "N/A") {
@@ -157,107 +80,7 @@ class MovieInfo extends React.Component<IMovieInfoProps> {
           aria-labelledby="reviewModal"
           aria-hidden="true"
         >
-          <div className="modal-dialog modal-dialog-centered" role="document">
-            <div className="modal-content">
-              <div className="modal-header">
-                <h3 className="modal-title" id="exampleModalLabel">
-                  REVIEW
-                </h3>
-                <button
-                  type="button"
-                  className="close"
-                  data-dismiss="modal"
-                  aria-label="Close"
-                >
-                  <span aria-hidden="true">&times;</span>
-                </button>
-              </div>
-              <div className="modal-body">
-                <form>
-                  <div className="form-group">
-                    <label>
-                      <strong>Review</strong>
-                    </label>
-                    <textarea
-                      className="form-control"
-                      id="exampleFormControlTextarea1"
-                      rows={3}
-                    />
-                  </div>
-                  <div className="form-group">
-                    <label>
-                      <strong>Rating</strong>
-                    </label>
-                    <div className="form-row">
-                      {/* 1 star */}
-                      <div className="form-check form-check-inline">
-                        <input
-                          className="form-check-input"
-                          type="radio"
-                          name="inlineRadioOptions"
-                          id="inlineRadio1"
-                          value="option1"
-                        />
-                        <label className="form-check-label">1</label>
-                      </div>
-                      {/* 2 star */}
-                      <div className="form-check form-check-inline">
-                        <input
-                          className="form-check-input"
-                          type="radio"
-                          name="inlineRadioOptions"
-                          id="inlineRadio2"
-                          value="option2"
-                        />
-                        <label className="form-check-label">2</label>
-                      </div>
-                      {/* 3 star */}
-                      <div className="form-check form-check-inline">
-                        <input
-                          className="form-check-input"
-                          type="radio"
-                          name="inlineRadioOptions"
-                          id="inlineRadio3"
-                          value="option3"
-                        />
-                        <label className="form-check-label">3</label>
-                      </div>
-                      {/* 4 star */}
-                      <div className="form-check form-check-inline">
-                        <input
-                          className="form-check-input"
-                          type="radio"
-                          name="inlineRadioOptions"
-                          id="inlineRadio4"
-                          value="option4"
-                        />
-                        <label className="form-check-label">4</label>
-                      </div>
-                      {/* 5 star */}
-                      <div className="form-check form-check-inline">
-                        <input
-                          className="form-check-input"
-                          type="radio"
-                          name="inlineRadioOptions"
-                          id="inlineRadio5"
-                          value="option5"
-                        />
-                        <label className="form-check-label">5</label>
-                      </div>
-                    </div>
-                  </div>
-                </form>
-                <button
-                  type="button"
-                  className="btn btn-block"
-                  id="rb-btn"
-                  onClick={() => this.loginUser()}
-                >
-                  Post Review
-                </button>
-              </div>
-            </div>
-          </div>
+          <Review />
         </div>
 
         {/* signup modal */}
@@ -268,102 +91,7 @@ class MovieInfo extends React.Component<IMovieInfoProps> {
           aria-labelledby="signupModal"
           aria-hidden="true"
         >
-          <div className="modal-dialog modal-dialog-centered" role="document">
-            <div className="modal-content">
-              <div className="modal-header">
-                <h3 className="modal-title" id="exampleModalLabel">
-                  SIGNUP
-                </h3>
-                <button
-                  type="button"
-                  className="close"
-                  data-dismiss="modal"
-                  aria-label="Close"
-                >
-                  <span aria-hidden="true">&times;</span>
-                </button>
-              </div>
-              <div className="modal-body">
-                <form>
-                  <div className="form-row">
-                    <div className="form-group col-6">
-                      <label>
-                        <strong>First Name</strong>
-                      </label>
-                      <input
-                        type="text"
-                        className="form-control"
-                        onChange={e => this.handleSignupFirstnameUpdate(e)}
-                      />
-                    </div>
-                    <div className="form-group col-6">
-                      <label>
-                        <strong>Last Name</strong>
-                      </label>
-                      <input
-                        type="text"
-                        className="form-control"
-                        onChange={e => this.handleSignupLastnameUpdate(e)}
-                      />
-                    </div>
-                  </div>
-                  <div className="form-group">
-                    <label>
-                      <strong>Email</strong>
-                    </label>
-                    <input
-                      type="email"
-                      className="form-control"
-                      onChange={e => this.handleSignupEmailUpdate(e)}
-                    />
-                  </div>
-                  <div className="form-row">
-                    <div className="form-group col-6">
-                      <label>
-                        <strong>Password</strong>
-                      </label>
-                      <input
-                        type="password"
-                        className="form-control"
-                        onChange={e => this.handleSignupPasswordUpdate(e)}
-                      />
-                    </div>
-                    <div className="form-group col-6">
-                      <label>
-                        <strong>Confirm Password</strong>
-                      </label>
-                      <input
-                        type="password"
-                        className="form-control"
-                        onChange={e => this.handleSignupConfirmPassword(e)}
-                      />
-                    </div>
-                  </div>
-                </form>
-                <button
-                  type="button"
-                  className="btn btn-block"
-                  id="rb-btn"
-                  onClick={() => this.signupUser()}
-                >
-                  Signup
-                </button>
-              </div>
-              <div className="modal-footer">
-                {"Already have an account? "}
-                <button
-                  className="signup btn"
-                  data-dismiss="modal"
-                  id="rb-btn"
-                  data-toggle="modal"
-                  data-target="#login-modal"
-                >
-                  {" "}
-                  login here
-                </button>
-              </div>
-            </div>
-          </div>
+          <Signup />
         </div>
 
         {/* login modal */}
@@ -374,68 +102,7 @@ class MovieInfo extends React.Component<IMovieInfoProps> {
           aria-labelledby="loginModal"
           aria-hidden="true"
         >
-          <div className="modal-dialog modal-dialog-centered" role="document">
-            <div className="modal-content">
-              <div className="modal-header">
-                <h3 className="modal-title" id="exampleModalLabel">
-                  LOGIN
-                </h3>
-                <button
-                  type="button"
-                  className="close"
-                  data-dismiss="modal"
-                  aria-label="Close"
-                >
-                  <span aria-hidden="true">&times;</span>
-                </button>
-              </div>
-              <div className="modal-body">
-                <form>
-                  <div className="form-group">
-                    <label>
-                      <strong>Email address</strong>
-                    </label>
-                    <input
-                      type="text"
-                      className="form-control"
-                      onChange={e => this.handleLoginEmailUpdate(e)}
-                    />
-                  </div>
-                  <div className="form-group">
-                    <label>
-                      <strong>Password</strong>
-                    </label>
-                    <input
-                      type="password"
-                      className="form-control"
-                      onChange={e => this.handleLoginPasswordUpdate(e)}
-                    />
-                  </div>
-                </form>
-                <button
-                  type="button"
-                  className="btn btn-block"
-                  id="rb-btn"
-                  onClick={() => this.loginUser()}
-                >
-                  Login
-                </button>
-              </div>
-              <div className="modal-footer">
-                Don't have an account?
-                <button
-                  className="signup btn"
-                  id="rb-btn"
-                  data-dismiss="modal"
-                  data-toggle="modal"
-                  data-target="#signup-modal"
-                >
-                  {" "}
-                  Sign up here
-                </button>
-              </div>
-            </div>
-          </div>
+          <Login />
         </div>
       </div>
     );
@@ -443,18 +110,10 @@ class MovieInfo extends React.Component<IMovieInfoProps> {
 }
 
 const mapStateToProps = (state: IState) => ({
-  login: state.login,
   movie: state.movie
 });
 
 const mapDispatchToProps = {
-  loginEmailUpdate: loginEmailUpdate,
-  loginPasswordUpdate: loginPasswordUpdate,
-  signupEmailUpdate: signupEmailUpdate,
-  signupPasswordUpdate: signupPasswordUpdate,
-  signupConfirmPassword: signupConfirmPassword,
-  signupFirstnameUpdate: signupFirstnameUpdate,
-  signupLastnameUpdate: signupLastnameUpdate,
   movieSearchResolved: movieSearchResolved,
   userSubmitRequest: userSubmitRequest,
   inputUpdate: inputUpdate,
