+++ client/src/containers/index.tsx
@@ -1,5 +1,7 @@
-export { default as MovieComponent } from './partials/movie.component';
-export { default as DashboardPg } from './pgs/Dashboard';
-export { default as SomethingWentWrongPg } from './pgs/404';
-export { default as SingleReviewPg } from './pgs/SingleReview';
-export { default as SearchResults } from './pgs/SearchResults';
+export { default as MovieComponent } from "./partials/movie.component";
+export { default as DashboardPg } from "./pgs/Dashboard";
+export { default as SomethingWentWrongPg } from "./pgs/404";
+export { default as SingleReviewPg } from "./pgs/SingleReview";
+export { default as SearchResults } from "./pgs/SearchResults";
+export { default as HomePg } from "./pgs/HomePg";
+export { default as ProfilePg } from "./pgs/ProfilePg";
